Rename id of orientation button to `orientation-button` (#10)
In Firefox and possibly some versions of Chrome, having an element with an id of `orientation` replaced the `window.orientation` numerical value with a pointer to the element. This appears to be fixed in upcoming versions of Firefox, but I'm renaming it anyway to fix it in the current version and any other browsers where it may be an issue.

Orientation in Firefox is still wonky, but that's a separate matter.